Complete code with Wayne stuff added and figure adjustments. Also is compatible with the old variance_offline data. Some variations of figures can be generated by commenting/uncommenting code (stripplot, error bars, horizontal lines instead of bars, etc)

+++ utils/offline_data.py
@@ -121,7 +121,7 @@ def split_offline_data(data, numFolds):
         for feat in data.keys():
             fulldat = data[feat]
             #for each context and each feat, add the leave in group for this fold.
-            newdata[feat].append(fulldat[inidx, ...])
+            newdata[feat].append(fulldat[inidx.astype(int), ...])
 
     return newdata, inIDXList, outIDXList
